[fact] Refactor SpoonNode into an interface #26
